+++ test/integration/033_event_tracking_test/test_events.py
@@ -105,6 +105,24 @@ class TestEventTracking(DBTIntegrationTest):
             }]
         return populate
 
+    def resource_counts_context(self):
+       return [
+          {
+              'schema': 'iglu:com.dbt/resource_counts/jsonschema/1-0-0',
+              'data': {
+                  'models': ANY,
+                  'tests': ANY,
+                  'snapshots': ANY,
+                  'analyses': ANY,
+                  'macros': ANY,
+                  'operations': ANY,
+                  'seeds': ANY,
+                  'sources': ANY,
+                  'exposures': ANY,
+              }
+          }
+       ]
+
     def build_context(
         self,
         command,
@@ -233,6 +251,12 @@ class TestEventTrackingSuccess(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='invocation',
                 label='end',
                 context=ANY
@@ -242,6 +266,7 @@ class TestEventTrackingSuccess(TestEventTracking):
         expected_contexts = [
             self.build_context('compile', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.build_context('compile', 'end', result_type='ok')
         ]
 
@@ -347,6 +372,12 @@ class TestEventTrackingSuccess(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='run_model',
                 label=ANY,
                 context=ANY
@@ -362,6 +393,7 @@ class TestEventTrackingSuccess(TestEventTracking):
         expected_contexts = [
             self.build_context('seed', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             seed_context,
             self.build_context('seed', 'end', result_type='ok')
         ]
@@ -380,6 +412,12 @@ class TestEventTrackingSuccess(TestEventTracking):
             call(
                 category='dbt',
                 action='load_project',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
+                action='resource_counts',
                 label=ANY,
                 context=ANY
             ),
@@ -412,6 +450,7 @@ class TestEventTrackingSuccess(TestEventTracking):
         expected_contexts = [
             self.build_context('run', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.run_context(
                 hashed_contents='1e5789d34cddfbd5da47d7713aa9191c',
                 model_id='4fbacae0e1b69924b22964b457148fb8',
@@ -457,6 +496,12 @@ class TestEventTrackingSuccess(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='run_model',
                 label=ANY,
                 context=ANY
@@ -472,6 +517,7 @@ class TestEventTrackingSuccess(TestEventTracking):
         expected_contexts = [
             self.build_context('run', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.run_context(
                 hashed_contents='4419e809ce0995d99026299e54266037',
                 model_id='576c3d4489593f00fad42b97c278641e',
@@ -510,6 +556,12 @@ class TestEventTrackingSuccess(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='invocation',
                 label='end',
                 context=ANY
@@ -519,6 +571,7 @@ class TestEventTrackingSuccess(TestEventTracking):
         expected_contexts = [
             self.build_context('test', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.build_context('test', 'end', result_type='ok')
         ]
 
@@ -621,6 +674,12 @@ class TestEventTrackingUnableToConnect(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='invocation',
                 label='end',
                 context=ANY
@@ -630,6 +689,7 @@ class TestEventTrackingUnableToConnect(TestEventTracking):
         expected_contexts = [
             self.build_context('run', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.build_context('run', 'end', result_type='error')
         ]
 
@@ -668,6 +728,12 @@ class TestEventTrackingSnapshot(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY
+            ),
+            call(
+                category='dbt',
                 action='run_model',
                 label=ANY,
                 context=ANY
@@ -684,6 +750,7 @@ class TestEventTrackingSnapshot(TestEventTracking):
         expected_contexts = [
             self.build_context('snapshot', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.run_context(
                 hashed_contents=ANY,
                 model_id='820793a4def8d8a38d109a9709374849',
@@ -723,6 +790,12 @@ class TestEventTrackingCatalogGenerate(TestEventTracking):
             ),
             call(
                 category='dbt',
+                action='resource_counts',
+                label=ANY,
+                context=ANY,
+            ),
+            call(
+                category='dbt',
                 action='invocation',
                 label='end',
                 context=ANY
@@ -732,6 +805,7 @@ class TestEventTrackingCatalogGenerate(TestEventTracking):
         expected_contexts = [
             self.build_context('generate', 'start'),
             self.load_context(),
+            self.resource_counts_context(),
             self.build_context('generate', 'end', result_type='ok')
         ]